Adds content tests for .html files
