+++ src/content/en/_shared/feedback.html
@@ -1,3 +1,4 @@
+{# wf_uses_template #}
 {% comment %}
   See https://web-central.appspot.com/web/resources/widgets#inline-feedback for
   usage.
