+++ src/content/en/ilt/pwa/_slides-slides-only.html
@@ -1,3 +1,4 @@
+{# wf_uses_template #}
 <style>
   .ilt-slides-wrapper iframe {
     height: 100%;
