+++ src/content/en/ilt/pwa/_slides.html
@@ -1,3 +1,4 @@
+{# wf_uses_template #}
 <style>
   .ilt-video-wrapper {
     float: left;
