+++ src/content/en/tools/chrome-devtools/javascript/_feedback/6.html
@@ -1,3 +1,4 @@
+{# wf_uses_template #}
 {% setvar category "DevTools" %}
 {% setvar label "JS / Get Started / (6) Demo Is Fixed" %}
 {% setvar url %}https://github.com/google/webfundamentals/issues/new?title=[{{ category }} / {{ label }}] {% endsetvar %}
