+++ src/content/en/tools/chrome-devtools/javascript/_feedback/7.html
@@ -1,3 +1,4 @@
+{# wf_uses_template #}
 {% setvar question "Was this tutorial helpful?" %}
 {% setvar category "DevTools" %}
 {% setvar label "JS / Get Started / Tutorial Is Helpful" %}
