+++ src/content/en/tools/chrome-devtools/javascript/_feedback/8.html
@@ -1,3 +1,4 @@
+{# wf_uses_template #}
 {% setvar question "Was this tutorial too long?" %}
 {% setvar category "DevTools" %}
 {% setvar label "JS / Get Started / Tutorial Is Too Long" %}
