+++ src/content/en/tools/lighthouse/audits/_feedback/_helpful.html
@@ -1,3 +1,4 @@
+{# wf_uses_template #}
 {% setvar category "Lighthouse" %}
 {% setvar question "Was this page helpful?" %}
 {% setvar url %}https://github.com/google/webfundamentals/issues/new?title=[{{ category }} / {{ label }}] {% endsetvar %}
